Move api config
